# trade-scheduler/backend/src/api-zod/generated/types/errorResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Scheduler 2 API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ErrorResponse {
  error: string;
  details?: string | null;
}