# trade-scheduler/backend/src/api-zod/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Scheduler 2 API specification
 * OpenAPI spec version: 0.2.0
 */
import type { JobPriority } from "./jobPriority";
import type { JobStatus } from "./jobStatus";
import type { JobType } from "./jobType";
import type { ValidityCode } from "./validityCode";
import type { Worker } from "./worker";

export interface Job {
  id: number;
  jobType: JobType;
  validityCode: ValidityCode;
  title: string;
  description?: string | null;
  clientName: string;
  clientPhone?: string | null;
  clientEmail?: string | null;
  address: string;
  latitude?: number | null;
  longitude?: number | null;
  price: number;
  estimatedHours: number;
  /** Number of tradies required */
  numTradies: number;
  /** Hours before job to call up for materials delivery */
  callUpTimeHours?: number | null;
  status: JobStatus;
  priority: JobPriority;
  /** Code 9 emergency override */
  isEmergency: boolean;
  scheduledDate?: string | null;
  completedDate?: string | null;
  notes?: string | null;
  tradeType: string;
  /** IDs of assigned workers */
  assignedWorkerIds: number[];
  /** Populated assigned workers */
  assignedWorkers: Worker[];
  distanceKm?: number | null;
  smartScore?: number | null;
  /** Estimated travel time in minutes */
  travelTimeMinutes?: number | null;
  invoiceNumber?: string | null;
  invoiceSentAt?: string | null;
  customerConfirmed: boolean;
  createdAt: string;
  updatedAt: string;
}