# trade-scheduler/backend/src/api-zod/generated/types/jobPriority.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Scheduler 2 API specification
 * OpenAPI spec version: 0.2.0
 */

export type JobPriority = (typeof JobPriority)[keyof typeof JobPriority];

export const JobPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;