# trade-scheduler/backend/src/api-zod/generated/types/updateJobRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Scheduler 2 API specification
 * OpenAPI spec version: 0.2.0
 */
import type { JobPriority } from "./jobPriority";
import type { JobStatus } from "./jobStatus";
import type { JobType } from "./jobType";
import type { ValidityCode } from "./validityCode";

export interface UpdateJobRequest {
  jobType?: JobType;
  validityCode?: ValidityCode;
  title?: string;
  description?: string | null;
  clientName?: string;
  clientPhone?: string | null;
  clientEmail?: string | null;
  address?: string;
  latitude?: number | null;
  longitude?: number | null;
  price?: number;
  estimatedHours?: number;
  numTradies?: number;
  callUpTimeHours?: number | null;
  status?: JobStatus;
  priority?: JobPriority;
  isEmergency?: boolean;
  scheduledDate?: string | null;
  completedDate?: string | null;
  notes?: string | null;
  tradeType?: string;
  assignedWorkerIds?: number[];
  customerConfirmed?: boolean;
  invoiceNumber?: string | null;
  invoiceSentAt?: string | null;
}