# trade-scheduler/backend/src/api-zod/generated/types/validityCode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Scheduler 2 API specification
 * OpenAPI spec version: 0.2.0
 */

/**
 * Job desirability rating: 1=low, 2=medium, 3=high priority
 */
export type ValidityCode = (typeof ValidityCode)[keyof typeof ValidityCode];

export const ValidityCode = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
} as const;